Samlar den generiske merge-to-pdfar-logikken i ei fil
